Changed Save implementation
The previous method of saving all files used the indirect method of
sending the Ctrl+Alt+S keystroke to Visual Studio.  While uncommon it is
possible for users or extensions to remap the Ctrl+Alt+S keystroke to a
different Visual Studio command.  If that happened the SaveAllTheTime
extension would begin executing that command instead of saving.

I switched the save implementation to use the more direct route of
actually executing the File.SaveAll command.  This is what Ctrl+Alt+S is
bound to by default.