Hook into project property designer
This change hooks the save logic into changes in the project which occur
through the project property designer (aka application designer).  This
is not a document based edit and hence doesn't have an IWpfTextView
associated with it.  Instead we need to hook into the lower level
interfaces of Visual Studio.  In particular IVsRunningDocumentTable and
IVsWindowFrameNotify

This fixes the first part of issue #5.  Still need to hook into the
adding / deleting of files in a project / solution