Fix MSB3270 warning on build
The SaveAllTheTime project references the
Microsoft.VisualStudio.Settings.11.0 assembly. This is actually a
C++/CLI project and hence is bound to x86.  The SaveAllTheTime project
was targeting MSIL and hence a MSB 3270 was being emitted on build.

This project can be safely changed to target the x86 platform.  Visual
Studio is a 32 bit only application hence this has no need to run on
x64.