Startup settings now associated with project
Previously the startup debugger settings were stored in the .csproj.user
file.  This meant that a new clone of SaveAllTheTime or branch switch
would eliminate the start up / debugging settings for the project.  By
putting them into the .csproj directly we eliminate this problem.

Additionally this uses the $(DevEnvDir) variable instead of hard coding
the directory to Visual Studio directory.  This makes the setting
tolerant to x86 versions of Windows as well as non-traditional install
locations (some people like to install Visual Studio on the D drive).